python/progressbar2/requirements.txt
progressbar2==3.6.0
wheel==0.24.0
